% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedBCGEData-package.R
\docType{package}
\name{AnnotatedBCGEData-package}
\alias{AnnotatedBCGEData}
\alias{AnnotatedBCGEData-package}
\title{AnnotatedBCGEData: 100+ Curated Breast Cancer Gene Expression Data sets}
\description{
AnnotatedBCGEData is an R Bioconductor package that provides access to more than 100 gene-expression datasets generated using microarray and RNA- sequencing technologies. We collected these datsets from public repositories, including Gene Expression Omnibus, ArrayExpress, and The Cancer Genome Atlas. Where feasible, we obtained and reprocessed the raw data using computational pipelines and annotations that were standardized and more updated than those used to process the data originally. Additionally, we annotated the sample metadata using oncology terms from the National Cancer Institute Thesaurus. This was necessary because different data submitters used different terminology to describe breast-cancer patients and samples. Mapping the metadata to ontology terms makes it possible for users to search the data using these standardized terms.
}
\section{Overview}{

AnnotatedBCGEData is an R Bioconductor package that provides access
to more than 100 gene-expression datasets generated using microarray and RNA-
sequencing technologies. We collected these datsets from public repositories,
including GEO, ArrayExpress, and TCGA. Where feasible, we obtained the and
reprocessed the raw data using computational pipelines and annotations that
were standardized and more updated than those used to process the data
originally. Additionally, we annotated the sample metadata using oncology
terms from the National Cancer Institute Thesaurus. This was necessary
because different data submitters used different terminology to describe
breast-cancer patients and samples. Mapping the metadata to ontology terms
makes it possible for users to search the data using these standardized terms.
\itemize{
\item \code{\link[=getDataset]{getDataset()}} allows users to get a SummarizedExperiment object of the chosen dataset.
\item \code{\link[=searchOntologyTerms]{searchOntologyTerms()}} allows users to browse our custom ontology.
\item \code{\link[=searchForDatasetsByField]{searchForDatasetsByField()}} is one function that finds datasets related to a given term.
}

Read the vignette for more information and detailed examples of the package functions.
}

\section{Examples}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{# Searching the ontology by a specific term name
searchOntologyTerms("Progesterone Receptor Status")

# Find associated datasets in the ontology mappings by a field code
searchForDatasetsByField("C16149")

# Retrieving a dataset
getDataset("GSE41197")
}\if{html}{\out{</div>}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/srp33/AnnotatedBCGEData}
  \item Report bugs at \url{https://github.com/srp33/AnnotatedBCGEData//issues}
}

}
\author{
\strong{Maintainer}: Stephen Piccolo \email{stephen_piccolo@byu.edu} (\href{https://orcid.org/0000-0003-2001-5640}{ORCID})

Authors:
\itemize{
  \item Heidi Steadman \email{hstead05@byu.edu}
  \item Shu-Fei Lo \email{shufeilo@byu.edu}
}

Other contributors:
\itemize{
  \item BYU College Undergraduate Research Award [funder]
}

}
\keyword{internal}
