% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasetMetadata.R
\name{getDatasetMetadata}
\alias{getDatasetMetadata}
\title{A function that takes a vector of dataset IDs and returns a tibble with
information about how the data was collected and the overall study.}
\usage{
getDatasetMetadata(datasets)
}
\arguments{
\item{datasets}{is a vector of strings of the dataset IDs to search for.}
}
\value{
Tibble with information about the experiment that produced the data.
}
\description{
A function that takes a vector of dataset IDs and returns a tibble with
information about how the data was collected and the overall study.
}
\examples{
getDatasetMetadata(c("GSE41197", "GSE10797"))
}
