% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\docType{data}
\name{identifiers}
\alias{identifiers}
\title{named list with the names as the data set accessions and values as a vector with c(expressions, metadata)}
\format{
a named list with data sets as names and a vector with URL identifiers as value
\describe{
\item{GSE41197}{c(data identifiers, metadata identifiers)}
\item{GSE10797}{c(data identifiers, metadata identifiers)}
\item{GSE59772}{c(data identifiers, metadata identifiers)}
}
}
\source{
generated internally for AnnotatedBCGEData
}
\usage{
identifiers
}
\description{
named list with the names as the data set accessions and values as a vector with c(expressions, metadata)
}
\keyword{datasets}
