% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{makeObject}
\alias{makeObject}
\title{Function that creates downloads data and creates SummarizedExperiment object for user to access}
\usage{
makeObject(identifiers)
}
\arguments{
\item{identifiers}{vector of identifiers found in identifier_list}
}
\value{
a SummarizedExperiment object for the data set
}
\description{
Takes a vector with the identifiers for the download URL in the form c(expression data, metadata)
}
