% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchForDatasetsByField.R
\name{searchForDatasetsByField}
\alias{searchForDatasetsByField}
\title{Function that searches breast cancer datasets based on ontology terms mapped to data fields.}
\usage{
searchForDatasetsByField(code)
}
\arguments{
\item{code}{is an ontology term code retrieved using the searchOntologyTerms function.}
}
\value{
a tibble providing information about any identified datasets.
}
\description{
Accepts an ontology term code and searches for datasets based on fields
(metadata columns) that have been mapped to that ontology term.
Returns a tibble with the dataset identifier, field name, and
code. Users can then pass the dataset identifiers to the getDataset
function.
}
\examples{
searchForDatasetsByField("C16149")
}
