% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchForDatasetValues.R
\name{searchForDatasetValues}
\alias{searchForDatasetValues}
\title{Function that searches breast cancer ontology values mapped to data sets.}
\usage{
searchForDatasetValues(code)
}
\arguments{
\item{code}{is the NCIT code that was found from searchOntologyTerms.}
}
\value{
a data frame from the searched file with data set metadata.
}
\description{
Takes a search code and searches across datasets to find those that are
mapped to that term.
Returns a data frame that shows the name of the data set, the term, and its
code. Users can then pass the names of the datasets to getDataset to
get the SummarizedExperiment objects.
}
\examples{
searchForDatasetValues("C15496")
}
