% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-NCIT-terms.R
\name{searchOntologyTerms}
\alias{searchOntologyTerms}
\title{Function that searches NCIT definitions file.}
\usage{
searchOntologyTerms(term, term_type = "Name")
}
\arguments{
\item{term}{string to search df by.}

\item{term_type}{type of term; Name, URI, Code, or Definition.}
}
\value{
a data frame with rows that match search term.
}
\description{
Takes a term to search by and the type of the term. This could be the name of
the term itself, the associated code, the URI, or the general definition.
The function searches across all terms and returns matches
to the search with the name of the term, the URI, the code, and the
definition.
}
\examples{
searchOntologyTerms("Progesterone Receptor Status")
}
