% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{bcgeData}
\alias{bcgeData}
\title{Function that creates a SummarizedExperiment object

Takes a name of a data set and an optional version of the data, passes
the identifier vector to get concept DOI for specified data set.}
\usage{
bcgeData(datasetID, identifier = identifiers, zen = zenodom, v = NULL)
}
\arguments{
\item{datasetID}{the name of the data set, passed as a string}

\item{identifier}{list of identifiers, defaults to list included}

\item{v}{version of data to download, default to most recent}
}
\value{
a SummarizedExperiment object for the data set
}
\description{
Function that creates a SummarizedExperiment object

Takes a name of a data set and an optional version of the data, passes
the identifier vector to get concept DOI for specified data set.
}
\examples{
bcgeData("GSE41197")
}
