% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-data.R
\name{getBCGEData}
\alias{getBCGEData}
\title{Function that creates a SummarizedExperiment object with breast cancer gene
expression data.

Takes a name of the desired data set as a string, an optional file path for
the data to be saved to, and an optional version number. The function
looks for the data either in the specified location or in a temporary
directory. If the data is not found, it is downloaded. Returns the desired
data set in the form of a SummarizedExperiment object.}
\usage{
getBCGEData(
  datasetID,
  cacheDirPath = tempdir(),
  identifier = identifiers,
  v = NULL
)
}
\arguments{
\item{datasetID}{the name of the data set, passed as a string.}

\item{cacheDirPath}{the optional file path to where the cache should be saved.}

\item{identifier}{list of identifiers, defaults to list included.}

\item{v}{version of data to download, default to most recent.}
}
\value{
a SummarizedExperiment object for the data set.
}
\description{
Function that creates a SummarizedExperiment object with breast cancer gene
expression data.

Takes a name of the desired data set as a string, an optional file path for
the data to be saved to, and an optional version number. The function
looks for the data either in the specified location or in a temporary
directory. If the data is not found, it is downloaded. Returns the desired
data set in the form of a SummarizedExperiment object.
}
