% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDataset}
\alias{getDataset}
\title{Function that retrieves a breast cancer gene expression dataset.
Accepts the identifier of a desired dataset, an optional directory path for
the data to be cached in, and an optional version number. The function
looks for the data in the cache. If the data is not found, it is downloaded
and cached. The data are packaged as SummarizedExperiment objects. If the
user specifies a cache directory, downloaded files are stored in this
location so they don't need to be re-downloaded. The default is to use the
R session's temporary directory, which is removed after the session ends.}
\usage{
getDataset(datasetID, cacheDirPath = tempdir(), v = NULL)
}
\arguments{
\item{datasetID}{The identifier of the data set.}

\item{cacheDirPath}{The optional directory path to where the cache should be saved.}

\item{v}{Version of data to download. Defaults to most recent.}
}
\value{
A SummarizedExperiment object for the dataset.
}
\description{
Function that retrieves a breast cancer gene expression dataset.
Accepts the identifier of a desired dataset, an optional directory path for
the data to be cached in, and an optional version number. The function
looks for the data in the cache. If the data is not found, it is downloaded
and cached. The data are packaged as SummarizedExperiment objects. If the
user specifies a cache directory, downloaded files are stored in this
location so they don't need to be re-downloaded. The default is to use the
R session's temporary directory, which is removed after the session ends.
}
\examples{
getDataset("GSE41197")
}
