% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasetMetadata.R
\name{getDatasetMetadata}
\alias{getDatasetMetadata}
\title{A function that accepts a vector of dataset IDs and returns a tibble with
metadata about the overall study.}
\usage{
getDatasetMetadata(datasets)
}
\arguments{
\item{datasets}{A vector of strings consisting of dataset identifier(s).}
}
\value{
Tibble with metadata.
}
\description{
A function that accepts a vector of dataset IDs and returns a tibble with
metadata about the overall study.
}
\examples{
getDatasetMetadata(c("GSE41197", "GSE10797"))
}
