% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-NCIT-terms.R
\name{searchDefs}
\alias{searchDefs}
\title{Function that searches NCIT definitions file}
\usage{
searchDefs(term, term_type = "Name")
}
\arguments{
\item{term}{string to search df by}

\item{term_type}{column to search in df, default Name}
}
\value{
a dataframe with matching rows to search term
}
\description{
Takes a search term and the column of the defs file to search by
Returns a dataframe with all matches to the parameters
}
