% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-mapped-ontology.R
\name{searchForDatasets}
\alias{searchForDatasets}
\title{Function that searches mapped ontology data}
\usage{
searchForDatasets(code, term_type = "field")
}
\arguments{
\item{code}{NCIT code that was found from searching defs}

\item{term_type}{field or value}
}
\value{
a data frame from the searched file with dataset metadata
}
\description{
Takes a search code and column to search by
Returns data frame joined with dataset metadata
}
\examples{
searchForDatasets("C19790", "field")
}
