% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchOntologyTerms.R
\name{searchOntologyTerms}
\alias{searchOntologyTerms}
\title{Function that searches NCIT definitions file.}
\usage{
searchOntologyTerms(term, term_type = "Name")
}
\arguments{
\item{String}{to search by.}

\item{Ontology}{field to search: Name, URI, Code, or Definition.}
}
\value{
A tibble with a row for each matching ontology term.
}
\description{
Takes a term to search by and the type of the term. This could be the name of
the term itself, the ontology's definition, the associated code, or the URI.
The function searches across all terms and returns matches
to the search with the term name, URI, code, and definition.
}
\examples{
searchOntologyTerms("Progesterone Receptor Status")
}
